pyproject.toml
[project]
name = "shreve-ff-dash"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nfl-data-py",
    "fastapi",
    "uvicorn",
    "requests",
    "python-dotenv>=1.2.1",
]
